analytics/dagster/requirements.txt
dagster
dagster-dbt
dagster-duckdb
dagster-embedded-elt
GeoAlchemy2==0.16.0
dbt-duckdb
duckdb
matplotlib
pandas
requests
dlt[duckdb]
psycopg2-binary
pydantic_settings
dlt[filesystem]
dlt[parquet]
s3fs==2024.12.0
boto3==1.35.91
thefuzz==0.22.1
psutil
